+++ NIMIO2024_TFG_PolMorenoCastelar/src/main/java/com/nimio2024/nimio2024_tfg_polmorenocastelar/application/CenterRestController.java
@@ -41,5 +41,10 @@ public class CenterRestController {
 
     }
 
+    @DeleteMapping("/{centerId}")
+    public void deleteCenter(@PathVariable Long centerId) {
+        centerController.deleteCenter(centerId);
+    }
+
 
 }