+++ NIMIO2024_TFG_PolMorenoCastelar/src/main/java/com/nimio2024/nimio2024_tfg_polmorenocastelar/controllers/CenterController.java
@@ -36,4 +36,8 @@ public class CenterController {
     public List<School> getSchoolsByCenter(Long centerId) {
         return centerService.getSchoolsByCenterId(centerId);
     }
+
+    public void deleteCenter(Long centerId) {
+        centerService.deleteCenter(centerId);
+    }
 }