+++ NIMIO2024_TFG_PolMorenoCastelar/changelog.md
@@ -18,3 +18,4 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - [NPP-10] 》CE2F - Obtain the list of schools of a center
 - [NPP-11] 》CE3F - Get all centers registered in the system
 - [NPP-12] 》CE4F - Edit a Center by ID
+- [NPP-13] 》CE5F - Delete a Center by ID